// client/src/api/models/runSourceDto.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * docXtractor
 * A comprehensive docXtractor System to extract text from documents.
 * OpenAPI spec version: 1.0
 */
import type { RunSourceDtoType } from "./runSourceDtoType";

export interface RunSourceDto {
	type: RunSourceDtoType;
	/** Display name for the source */
	name: string;
	/** URL if type is url */
	url?: string;
	/** File ID if type is file */
	fileId?: string;
}